requirements.txt
Flask==2.3.3 
requests==2.31.0 
beautifulsoup4==4.12.2 
lxml==4.9.3 
schedule==1.2.0 
psycopg2-binary==2.9.6